+++ src/main/java/com/example/qsproject/qsproject/services/CourseService.java
@@ -2,12 +2,16 @@ package com.example.qsproject.qsproject.services;
 
 import com.example.qsproject.qsproject.dtos.CourseDto;
 
+import java.util.List;
+
 // Interface com os métodos que terao de ser utilizados no UserServiceImpl
 public interface CourseService {
 
     CourseDto createCourse(CourseDto coursesDto);
 
-    CourseDto getCourseById(int id);
+    CourseDto getCourseById(long id);
 
-    CourseDto deleteCourseById(int id);
+    CourseDto deleteCourseById(long id);
+
+    //List<CourseDto> getAllCourses();
 }
